+++ test/ERC20SnapshotModuleCommon/ERC20SnapshotModuleCommonUnschedule.js
@@ -1,6 +1,8 @@
 const { time } = require('@nomicfoundation/hardhat-network-helpers')
 const { expect } = require('chai')
-const { SNAPSHOOTER_ROLE } = require('../utils')
+const {
+  expectUnauthorizedSnapshotAdmin
+} = require('./ERC20SnapshotModuleUtils/authorization')
 const {
   checkArraySnapshot
 } = require('./ERC20SnapshotModuleUtils/ERC20SnapshotModuleUtils')
@@ -150,16 +152,13 @@ function ERC20SnapshotModuleCommonUnschedule () {
       expect(snapshots.length).to.equal(1)
       expect(snapshots[0]).to.equal(SNAPSHOT_TIME)
       // Act
-      await expect(
+      await expectUnauthorizedSnapshotAdmin(
+        this,
         this.transferEngineMock
           .connect(this.address1)
-          .unscheduleSnapshotNotOptimized(SNAPSHOT_TIME)
+          .unscheduleSnapshotNotOptimized(SNAPSHOT_TIME),
+        this.address1
       )
-        .to.be.revertedWithCustomError(
-          this.transferEngineMock,
-          'AccessControlUnauthorizedAccount'
-        )
-        .withArgs(this.address1.address, SNAPSHOOTER_ROLE)
       // Assert
       snapshots = await this.transferEngineMock.getNextSnapshots()
       expect(snapshots.length).to.equal(1)
@@ -185,16 +184,13 @@ function ERC20SnapshotModuleCommonUnschedule () {
     })
 
     it('reverts when calling from non-admin', async function () {
-      await expect(
+      await expectUnauthorizedSnapshotAdmin(
+        this,
         this.transferEngineMock
           .connect(this.address1)
-          .unscheduleLastSnapshot(this.snapshotTime)
+          .unscheduleLastSnapshot(this.snapshotTime),
+        this.address1
       )
-        .to.be.revertedWithCustomError(
-          this.transferEngineMock,
-          'AccessControlUnauthorizedAccount'
-        )
-        .withArgs(this.address1.address, SNAPSHOOTER_ROLE)
     })
 
     it('reverts if no snapshot is scheduled', async function () {